requirements.txt
Click==7.0
Flask==1.1.1
Flask-SocketIO==4.2.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dotenv==0.10.3
python-engineio==3.11.1
python-socketio==4.4.0
six==1.13.0
Werkzeug==0.16.0